backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
Requests==2.32.3
psycopg2>=2.8,<3.0
